tasks: add RunGradlew() and StopGradlew() shortcuts
We need a way to run our Java code straight from vim, no going around in
different tmux panes and running commands, this takes valuable development
time. We are in vim land, this means we have to use shortcuts for everything.

Make `RunGradlew()` and `StopGradlew()` tasks triggered by a series of
shortcuts including the leader key to make this work.

Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>